Update project with tool accessibility, making new entry point and end points, make new methods in tool service and create new requests like ToolCreateRequest.java, ToolUpdateRequest.java.

+++ src/main/java/me/blackwater/tools/web/requests/ToolCreateRequest.java
@@ -10,6 +10,6 @@ public record ToolCreateRequest(
         @Positive
         int price,
         @PositiveOrZero
-        int shopId
+        long shopId
 ) {
 }
